fix: extract icon-size signals to break circular dependency
Move iconSize$, iconSizes$, setIconSize$ and IconSize type to a
separate icon-size.ts to prevent the circular import between
internal-connector-logos-setup.ts and internal-connector-logos.tsx
